+++ packages/core/content-releases/admin/src/components/AddReleaseDialog.tsx
@@ -7,6 +7,7 @@ import {
   TextInput,
   Typography,
 } from '@strapi/design-system';
+import { useNotification } from '@strapi/helper-plugin';
 import { Formik, Form } from 'formik';
 import { useIntl } from 'react-intl';
 import * as yup from 'yup';
@@ -21,11 +22,23 @@ const INITIAL_VALUES = {
 
 interface AddReleaseDialogProps {
   handleClose: () => void;
-  handleSubmit: () => void;
 }
 
-export const AddReleaseDialog = ({ handleClose, handleSubmit }: AddReleaseDialogProps) => {
+export const AddReleaseDialog = ({ handleClose }: AddReleaseDialogProps) => {
   const { formatMessage } = useIntl();
+  const toggleNotification = useNotification();
+
+  const handleSubmit = () => {
+    handleClose();
+
+    toggleNotification({
+      type: 'success',
+      message: formatMessage({
+        id: 'content-releases.modal.release-created-notification-success',
+        defaultMessage: 'Release created.',
+      }),
+    });
+  };
 
   return (
     <ModalLayout onClose={handleClose} labelledBy="title">
