+++ packages/core/content-releases/admin/src/pages/ReleasesPage/Releases.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 
 import { Button, HeaderLayout } from '@strapi/design-system';
-import { useNotification } from '@strapi/helper-plugin';
 import { Plus } from '@strapi/icons';
 import { useIntl } from 'react-intl';
 
@@ -10,24 +9,11 @@ import { AddReleaseDialog } from '../../components/AddReleaseDialog';
 const ReleasesPage = () => {
   const [addReleaseDialogIsShown, setAddReleaseDialogIsShown] = React.useState(false);
   const { formatMessage } = useIntl();
-  const toggleNotification = useNotification();
 
   const total = 0; // TODO: replace it with the total number of releases
 
   const toggleAddReleaseDialog = () => {
     setAddReleaseDialogIsShown((prev) => !prev);
-  };
-
-  const handleSubmit = () => {
-    toggleAddReleaseDialog();
-
-    toggleNotification({
-      type: 'success',
-      message: formatMessage({
-        id: 'content-releases.modal.release-created-notification-success',
-        defaultMessage: 'Release created.',
-      }),
-    });
   };
 
   return (
@@ -53,9 +39,7 @@ const ReleasesPage = () => {
           </Button>
         }
       />
-      {addReleaseDialogIsShown && (
-        <AddReleaseDialog handleClose={toggleAddReleaseDialog} handleSubmit={handleSubmit} />
-      )}
+      {addReleaseDialogIsShown && <AddReleaseDialog handleClose={toggleAddReleaseDialog} />}
     </>
   );
 };
